Enhance: Cleanup c/cpp code (#242)
Currently, the code style of miluvs-storage is not entirely consistent.
The current changes aim to unify the code style as much as possible and
modify some code that does not follow common C++ practices, including:

- Using get() on shared_ptr/unique_ptr to obtain a raw pointer for
method calls. Which is unnecessary and dangerous.
- Removed class-private unused variables and local unused variables.
- Implicit type conversions, ex `for (int i = 0; i < size_t)`, which can
be dangerous.
- Unnecessary `std::move`, such as using `std::move` in cases where
Return Value Optimization (RVO) already applies.
- Missing `std::move` (resulting in extra memory copies), such as when a
local variableis assigned and then passed to `push_back` or
`emplace_back`.
- Replacing most `push_back` calls with `emplace_back`, as `push_back`
always incurs an additional memory copy.
- Renaming some methods that used underscore naming to camelCase.

This commit does not yet cover all code comprehensively—only parts that
were encountered during code review have been modified.

+++ cpp/src/packed/chunk_manager.cpp
@@ -39,7 +39,7 @@ std::vector<std::shared_ptr<arrow::ArrayData>> ChunkManager::SliceChunksByMaxCon
   std::set<int> table_to_pop;
 
   // Identify the chunk for each column and adjust chunk size
-  for (int i = 0; i < column_offsets_.size(); ++i) {
+  for (size_t i = 0; i < column_offsets_.size(); ++i) {
     auto offset = column_offsets_[i];
     auto& table_queue = tables[offset.path_index];
 
@@ -68,7 +68,7 @@ std::vector<std::shared_ptr<arrow::ArrayData>> ChunkManager::SliceChunksByMaxCon
 
   // Slice chunks and advance chunk index as appropriate
   std::vector<std::shared_ptr<arrow::ArrayData>> batch_data(column_offsets_.size());
-  for (int i = 0; i < column_offsets_.size(); ++i) {
+  for (size_t i = 0; i < column_offsets_.size(); ++i) {
     auto& chunk_state = chunk_states_[i];
     const auto& chunk = chunks[i];
     int64_t offset = chunk_state.offset;
@@ -105,7 +105,7 @@ std::vector<std::shared_ptr<arrow::ArrayData>> ChunkManager::SliceChunksByMaxCon
 
 // resets the chunk states for columns in a specific file.
 void ChunkManager::ResetChunkState(int path_index) {
-  for (int j = 0; j < column_offsets_.size(); ++j) {
+  for (size_t j = 0; j < column_offsets_.size(); ++j) {
     if (column_offsets_[j].path_index == path_index) {
       chunk_states_[j].reset();
     }
